Implement collision & bounds checking for pieces
In similar vein to shapeEncoding, we define a 4x4 occupancy mask as a 16 bit
integer representing a 4x4 grid, with the coordinate (0,0) corresponding to the
LSB. Lit bits indicate a blocked position, either because it's occupied or
because it is out of bounds.

By ANDing the shapeEncoding of our piece with this mask, we can
immediately tell if there's a collision(the AND result is not 0).

Also implemented a Debug.displayShape function, which uses String functions
(fulfilling yet another exercise requirement, I might keep it in the submitted
exercise)